Refactor upload logic into separate modules
